feat(law): add Precept of the Scribed Weave to meta spellbook
Distills a learning from a previous ritual failure where a file change was not committed before merging. This new precept makes it Law that an act of shaping (edit_file) must be followed by an act of scribing (git commit) before a weaving rite (git merge) can begin. This strengthens the integrity of meta-practice.